# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/community/dashboard": {
      "get": {
        "operationId": "CommunityController_dashboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Comunidade"
        ]
      }
    },
    "/api/v1/community/visitors": {
      "get": {
        "operationId": "CommunityController_visitors",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100000,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 20,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Comunidade"
        ]
      },
      "post": {
        "operationId": "CommunityController_visitor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisitorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Comunidade"
        ]
      }
    },
    "/api/v1/community/visitors/{id}": {
      "patch": {
        "operationId": "CommunityController_updateVisitor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisitorPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Comunidade"
        ]
      }
    },
    "/api/v1/community/events": {
      "get": {
        "operationId": "CommunityController_events",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100000,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 20,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Comunidade"
        ]
      },
      "post": {
        "operationId": "CommunityController_event",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Comunidade"
        ]
      }
    },
    "/api/v1/community/events/{id}": {
      "patch": {
        "operationId": "CommunityController_updateEvent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Comunidade"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ponte API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "VisitorDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "pattern": "\\S"
          },
          "email": {
            "type": "string",
            "nullable": true,
            "maxLength": 320,
            "format": "email"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "maxLength": 32
          },
          "visitedAt": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          },
          "status": {
            "type": "string",
            "enum": [
              "NEW",
              "CONTACTED",
              "ARCHIVED"
            ]
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "maxLength": 2000
          }
        },
        "required": [
          "name",
          "visitedAt"
        ]
      },
      "VisitorPatch": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "pattern": "\\S"
          },
          "email": {
            "type": "string",
            "nullable": true,
            "maxLength": 320,
            "format": "email"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "maxLength": 32
          },
          "visitedAt": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          },
          "status": {
            "type": "string",
            "enum": [
              "NEW",
              "CONTACTED",
              "ARCHIVED"
            ]
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "maxLength": 2000
          }
        }
      },
      "EventDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 160,
            "pattern": "\\S"
          },
          "startsAt": {
            "type": "string"
          },
          "location": {
            "type": "string",
            "maxLength": 200,
            "pattern": "\\S"
          },
          "status": {
            "type": "string",
            "enum": [
              "SCHEDULED",
              "COMPLETED",
              "CANCELLED"
            ]
          },
          "description": {
            "type": "string",
            "nullable": true,
            "maxLength": 2000
          }
        },
        "required": [
          "name",
          "startsAt",
          "location"
        ]
      },
      "EventPatch": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 160,
            "pattern": "\\S"
          },
          "startsAt": {
            "type": "string"
          },
          "location": {
            "type": "string",
            "maxLength": 200,
            "pattern": "\\S"
          },
          "status": {
            "type": "string",
            "enum": [
              "SCHEDULED",
              "COMPLETED",
              "CANCELLED"
            ]
          },
          "description": {
            "type": "string",
            "nullable": true,
            "maxLength": 2000
          }
        }
      },
...
    }
  }
}